09.03.2026

+++ demo-hotel-service/src/main/java/com/demo_hotel_service/services/ServiceUnitService.java
@@ -1,10 +1,6 @@
 package com.demo_hotel_service.services;
 
-import java.io.IOException;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -36,7 +32,7 @@ public class ServiceUnitService {
     private final int DEFAULT_PAGE_SIZE = 12;
 
     public String getServiceUnitsHtml(Model model, String template, String attributeName,
-    Class<? extends ServiceUnit> classType) {
+            Class<? extends ServiceUnit> classType) {
         List<ServiceUnitDto> serviceUnitDtos = getServiceUnits(classType,
                 PageRequest.of(DEFAULT_PAGE_NUMBER, DEFAULT_PAGE_SIZE));
         model.addAttribute(attributeName, serviceUnitDtos);
@@ -69,11 +65,9 @@ public class ServiceUnitService {
         }
 
         var entity = serviceUnitMapper.toPolymorphicEntity(serviceUnitDto);
-        int i = 0;
         for (ImageRecord imageRecord : entity.getImageRecords()) {
             imageRecord.setServiceUnit(entity);
         }
-
 
         ServiceUnitDto savedDto = serviceUnitMapper
                 .toPolymorphicDto(serviceUnitRepository.save(entity));
@@ -81,9 +75,18 @@ public class ServiceUnitService {
         return savedDto;
     }
 
-    public ServiceUnitDto updateServiceUnit(ServiceUnitDto updatingData) {
+    @Transactional
+    public ServiceUnitDto updateServiceUnit(ServiceUnitDto updatingData, List<MultipartFile> imageFiles) {
+        if (updatingData == null) {
+            throw new IllegalArgumentException("Data can't be null");
+        }
+
         ServiceUnit toUpdate = serviceUnitRepository.findById(updatingData.getId())
                 .orElseThrow(() -> new EntityNotFoundException("Service not found"));
+
+        if (imageFiles != null && !imageFiles.isEmpty()) {
+            imageService.addImages(updatingData.getId(), imageFiles);
+        }
 
         serviceUnitMapper.updateEntity(updatingData, toUpdate);
         ServiceUnitDto updatedDto = serviceUnitMapper.toPolymorphicDto(serviceUnitRepository.save(toUpdate));
@@ -91,24 +94,27 @@ public class ServiceUnitService {
     }
 
     @Transactional
-    public boolean deleteServiceUnit(long serviceId) throws IOException {
+    public Boolean deleteServiceUnit(Long serviceId) {
+        if (serviceId == null) {
+            throw new IllegalArgumentException();
+        }
         if (!serviceUnitRepository.existsById(serviceId)) {
             throw new EntityNotFoundException("Service not found");
         }
+
         imageService.removeImages(serviceId);
         serviceUnitRepository.deleteById(serviceId);
         return serviceUnitRepository.existsById(serviceId);
     }
 
-    private void sortImageRecordsOfServiceUnits(List<ServiceUnitDto> serviceUnitDtos){
-        serviceUnitDtos.forEach( su ->{
-            su.getImageRecords().sort((img1, img2)-> img1.getPosition() > img2.getPosition() ? 1 : -1);
+    private void sortImageRecordsOfServiceUnits(List<ServiceUnitDto> serviceUnitDtos) {
+        serviceUnitDtos.forEach(su -> {
+            su.getImageRecords().sort((img1, img2) -> img1.getPosition() > img2.getPosition() ? 1 : -1);
         });
     }
 
-    private void sortImageRecordsOfServiceUnit(ServiceUnitDto serviceUnitDto){
+    private void sortImageRecordsOfServiceUnit(ServiceUnitDto serviceUnitDto) {
         serviceUnitDto.getImageRecords().sort((img1, img2) -> img1.getPosition() < img2.getPosition() ? 1 : -1);
     }
-
 
 }
